Extract common parallelism
